[refactor/#1] screen 패키지 구조 리팩토링
- Activity와 Screen 파일 분리
- 화면 유형 별로 패키지 구분
