+++ app/src/main/java/org/sopt/at/ui/screens/login/LoginScreen.kt
@@ -1,15 +1,6 @@
-@file:Suppress("UNCHECKED_CAST")
-
-package org.sopt.at
+package org.sopt.at.ui.screens.login
 
 import BackIcon
-import android.content.Intent
-import android.os.Build
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -29,16 +20,12 @@ import androidx.compose.material3.Button
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -54,86 +41,10 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.em
 import androidx.compose.ui.unit.sp
-import kotlinx.coroutines.launch
-import org.sopt.at.model.LoginUser
+import org.sopt.at.R
 import org.sopt.at.ui.theme.ATSOPTANDROIDTheme
 import org.sopt.at.ui.theme.ButtonDisableText
 import org.sopt.at.ui.theme.GuideText
-
-
-class LoginActivity : ComponentActivity() {
-    var loginUser: LoginUser? = null
-
-    private val signupResultLauncher = registerForActivityResult(
-        ActivityResultContracts.StartActivityForResult()
-    ) { result ->
-        if ( result.resultCode == RESULT_OK ) {
-            loginUser = (if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                result.data?.getParcelableExtra(SIGNUP_USER_INFO_KEY, LoginUser::class.java)
-            } else {
-                result.data?.getParcelableExtra<LoginUser>(SIGNUP_USER_INFO_KEY)
-            })
-        }
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            var idText by remember { mutableStateOf("") }
-            var pwdText by remember { mutableStateOf("") }
-
-            val scope = rememberCoroutineScope()
-            val snackbarHostState = remember { SnackbarHostState() }
-
-            ATSOPTANDROIDTheme {
-                Scaffold(
-                    modifier = Modifier.fillMaxSize(),
-                    snackbarHost = {
-                        SnackbarHost(hostState = snackbarHostState)
-                    },
-                ) { innerPadding ->
-                    LoginScreen(
-                        modifier = Modifier.padding(innerPadding),
-                        idText = idText,
-                        onIdChange = { idText = it },
-                        pwdText = pwdText,
-                        onPwdChange = { pwdText = it },
-                        onClickLoginButton = {
-                            if (isIdenticalLoginUserInfo(idText, pwdText)) {
-                                startActivity(Intent(this, MyActivity::class.java)
-                                    .putExtra(USER_ID_KEY, loginUser!!.id)
-                                )
-                            } else {
-                                val message = if (loginUser == null) {
-                                    "회원 정보가 없습니다."
-                                } else {
-                                    "아이디 또는 비밀번호가 일치하지 않습니다."
-                                }
-                                scope.launch {
-                                    snackbarHostState.showSnackbar(message)
-                                }
-                            }
-                        },
-                        onClickSignUpButton = {
-                            val intent = Intent(this, SignupActivity::class.java)
-                            signupResultLauncher.launch(intent)
-                        }
-                    )
-                }
-            }
-        }
-    }
-
-    fun isIdenticalLoginUserInfo(id: String, pwd: String): Boolean {
-        return loginUser?.id == id && pwd == loginUser?.password
-    }
-
-    companion object {
-        const val SIGNUP_USER_INFO_KEY = "signup_user_info_key"
-        const val USER_ID_KEY = "user_id_key"
-    }
-}
 
 @Composable
 fun LoginScreen(
@@ -148,7 +59,8 @@ fun LoginScreen(
     var isButtonEnable by remember { mutableStateOf(false) }
     var isPwdVisible by remember { mutableStateOf(false) }
 
-    val pwdIcon = if (isPwdVisible) painterResource(R.drawable.ic_password_show) else painterResource(R.drawable.ic_password_hide)
+    val pwdIcon = if (isPwdVisible) painterResource(R.drawable.ic_password_show) else painterResource(
+        R.drawable.ic_password_hide)
 
     fun updateButtonState() {
         isButtonEnable = idText.isNotEmpty() && pwdText.isNotEmpty()
